api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
pillow==11.2.1
faiss-cpu==1.11.0
pytest==8.3.5
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.3
email-validator==2.2.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.22.0
pytesseract>=0.3.10
pillow>=10.0.0 